renamed multiplatform package
