+++ appyx-components/backstack/common/src/commonMain/kotlin/com/bumble/appyx/components/backstack/operation/Pop.kt
@@ -5,7 +5,7 @@ import com.bumble.appyx.components.backstack.BackStack
 import com.bumble.appyx.components.backstack.BackStackModel.State
 import com.bumble.appyx.interactions.core.model.transition.BaseOperation
 import com.bumble.appyx.interactions.core.model.transition.Operation
-import com.bumble.appyx.utils.multiplatform_common.Parcelize
+import com.bumble.appyx.utils.multiplatform.Parcelize
 
 /**
  * Operation:
