+++ appyx-components/backstack/common/src/commonMain/kotlin/com/bumble/appyx/components/backstack/operation/Push.kt
@@ -6,8 +6,8 @@ import com.bumble.appyx.components.backstack.BackStackModel
 import com.bumble.appyx.interactions.core.asElement
 import com.bumble.appyx.interactions.core.model.transition.BaseOperation
 import com.bumble.appyx.interactions.core.model.transition.Operation
-import com.bumble.appyx.utils.multiplatform_common.Parcelize
-import com.bumble.appyx.utils.multiplatform_common.RawValue
+import com.bumble.appyx.utils.multiplatform.Parcelize
+import com.bumble.appyx.utils.multiplatform.RawValue
 
 /**
  * Operation:
