+++ appyx-components/demos/common/src/commonMain/kotlin/com/bumble/appyx/components/demos/promoter/operation/AddFirst.kt
@@ -6,8 +6,8 @@ import com.bumble.appyx.components.demos.promoter.PromoterModel
 import com.bumble.appyx.interactions.core.asElement
 import com.bumble.appyx.interactions.core.model.transition.BaseOperation
 import com.bumble.appyx.interactions.core.model.transition.Operation
-import com.bumble.appyx.utils.multiplatform_common.Parcelize
-import com.bumble.appyx.utils.multiplatform_common.RawValue
+import com.bumble.appyx.utils.multiplatform.Parcelize
+import com.bumble.appyx.utils.multiplatform.RawValue
 
 @Parcelize
 data class AddFirst<InteractionTarget>(
