+++ appyx-components/internal/common/src/commonMain/kotlin/com/bumble/appyx/components/internal/testdrive/operation/MoveTo.kt
@@ -5,7 +5,7 @@ import com.bumble.appyx.components.internal.testdrive.TestDrive
 import com.bumble.appyx.components.internal.testdrive.TestDriveModel
 import com.bumble.appyx.interactions.core.model.transition.BaseOperation
 import com.bumble.appyx.interactions.core.model.transition.Operation
-import com.bumble.appyx.utils.multiplatform_common.Parcelize
+import com.bumble.appyx.utils.multiplatform.Parcelize
 
 @Parcelize
 data class MoveTo<InteractionTarget>(
