+++ appyx-components/spotlight/common/src/commonMain/kotlin/com/bumble/appyx/components/spotlight/operation/First.kt
@@ -5,7 +5,7 @@ import com.bumble.appyx.components.spotlight.Spotlight
 import com.bumble.appyx.components.spotlight.SpotlightModel
 import com.bumble.appyx.interactions.core.model.transition.BaseOperation
 import com.bumble.appyx.interactions.core.model.transition.Operation
-import com.bumble.appyx.utils.multiplatform_common.Parcelize
+import com.bumble.appyx.utils.multiplatform.Parcelize
 
 
 @Parcelize
