+++ utils/multiplatform-common/src/androidMain/kotlin/com/bumble/appyx/utils/multiplatform/Parcelize.kt
@@ -1,4 +1,4 @@
-package com.bumble.appyx.utils.multiplatform_common
+package com.bumble.appyx.utils.multiplatform
 
 actual typealias Parcelize = kotlinx.parcelize.Parcelize
